# idp-server/api/openapi.yaml
openapi: 3.0.3
info:
  title: IDP Server API
  version: 0.2.0
  description: |
    OpenAPI contract for the current Go IdP / OAuth2 / OIDC server.
    This document is aligned with the implemented routes and handlers under
    `internal/interfaces/http`.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /login:
    get:
      tags: [authn]
      summary: Login page, login metadata, or federated OIDC callback
      description: |
        When called without `code`/`state`, this endpoint returns either HTML
        or JSON login metadata depending on the `Accept` header. When called
        with federated OIDC callback parameters, it may authenticate the user
        and redirect to the original `return_to` path.
      operationId: getLogin
      parameters:
        - name: method
          in: query
          required: false
          schema:
            type: string
            enum: [password, federated_oidc]
        - name: return_to
          in: query
          required: false
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: nonce
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Login page or JSON login metadata
          headers:
            Set-Cookie:
              schema:
                type: string
              description: May set `idp_csrf_token` and, after successful authentication, `idp_session`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginInfoResponse'
            text/html:
              schema:
                type: string
        '302':
          description: Redirect after successful login or federated OIDC flow
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              schema:
                type: string
        '400':
          description: Invalid request or invalid `return_to`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: User locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Login failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [authn]
      summary: Login with password or start federated OIDC
      description: |
        Submit username/password for local login, or set `method=federated_oidc`
        to start external OIDC login. The CSRF token may be provided in the
        body as `csrf_token` or in the `X-CSRF-Token` header, and must match
        the `idp_csrf_token` cookie set by `GET /login`.
      operationId: login
      parameters:
        - $ref: '#/components/parameters/XCSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful and returns session information
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Sets the `idp_session` cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
            text/html:
              schema:
                type: string
        '302':
          description: Redirect to `return_to`, `/login`, or upstream OIDC provider
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              schema:
                type: string
        '400':
          description: Invalid login request or unsupported authentication method
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User disabled or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: User locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Login rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Login failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /connect/logout:
    get:
      tags: [oidc]
      summary: Browser-facing RP initiated logout entrypoint
      description: |
        Prepares an IdP logout request inside the IdP browser context. When
        `client_id` and `post_logout_redirect_uri` are supplied, the redirect
        target must be pre-registered in `post_logout_redirect_uris`.
      operationId: getEndSession
      parameters:
        - name: client_id
          in: query
          required: false
          schema:
            type: string
        - name: post_logout_redirect_uri
          in: query
          required: false
          schema:
            type: string
            format: uri
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Logout page metadata
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Sets the `idp_csrf_token` cookie when needed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndSessionInfoResponse'
        '400':
          description: Invalid end session request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to generate CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [oidc]
      summary: End current IdP browser session and redirect back to the client
      description: |
        Clears the current `idp_session` in the IdP browser context. This is
        the browser-safe logout flow for relying parties. The CSRF token may be
        provided in the form body as `csrf_token` or in the `X-CSRF-Token`
        header, and must match the `idp_csrf_token` cookie.
      operationId: endSession
      parameters:
        - $ref: '#/components/parameters/XCSRFToken'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndSessionRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EndSessionRequest'
      responses:
        '200':
          description: Session ended without redirect
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Clears the `idp_session` cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '302':
          description: Session ended and browser redirected to the whitelisted post logout target
          headers:
            Location:
              schema:
                type: string
                format: uri
            Set-Cookie:
              schema:
                type: string
              description: Clears the `idp_session` cookie
        '400':
          description: Invalid client id or post logout redirect URI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: CSRF validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Logout failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /oauth2/clients:
    post:
      tags: [authz]
      summary: Create an OAuth client
      description: |
        Creates a new OAuth client and writes its core configuration, grant
        types, auth methods, scopes, and optional redirect URIs in one step.
      operationId: createClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: Client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateClientResponse'
        '400':
          description: Invalid client payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Client ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Client creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XCSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: |
        Optional CSRF token header. If omitted, the same token may be supplied
        in the request body as `csrf_token`. In both cases it must match the
        `idp_csrf_token` cookie.
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid request
...
    LoginInfoResponse:
      type: object
      required:
        - endpoint
        - message
        - csrf_token
        - federated_oidc_enabled
        - federated_oidc_endpoint
        - federated_oidc_method
      properties:
        endpoint:
          type: string
          example: login
        message:
          type: string
          example: submit username and password to login
        csrf_token:
          type: string
        return_to:
          type: string
          nullable: true
        federated_oidc_enabled:
          type: boolean
        federated_oidc_endpoint:
          type: string
          example: /login
        federated_oidc_method:
          type: string
          example: POST
    LoginRequest:
      type: object
      properties:
        method:
          type: string
          enum: [password, federated_oidc]
        username:
          type: string
        password:
          type: string
          format: password
        csrf_token:
          type: string
        return_to:
          type: string
        redirect_uri:
          type: string
        state:
          type: string
        code:
          type: string
        nonce:
          type: string
    LoginResponse:
      type: object
      required: [session_id, user_id, subject, expires_at]
      properties:
        session_id:
          type: string
          format: uuid
        user_id:
          type: integer
          format: int64
        subject:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
...
    LogoutResponse:
      type: object
      required: [logged_out]
      properties:
        logged_out:
          type: boolean
          example: true
    EndSessionRequest:
      type: object
      properties:
        client_id:
          type: string
        post_logout_redirect_uri:
          type: string
          format: uri
        state:
          type: string
        csrf_token:
          type: string
    EndSessionInfoResponse:
      type: object
      required: [endpoint, action, csrf_token, message]
      properties:
        endpoint:
          type: string
          example: end_session
        action:
          type: string
          example: /connect/logout
        client_id:
          type: string
        post_logout_redirect_uri:
          type: string
          format: uri
        state:
          type: string
        csrf_token:
          type: string
        message:
          type: string
          example: submit POST /connect/logout from the browser context to end the IdP session
...
    CreateClientRequest:
      type: object
      required:
        - client_id
        - client_name
        - grant_types
        - scopes
      properties:
        client_id:
          type: string
          example: demo-web-client
        client_name:
          type: string
          example: Demo Web Client
        client_secret:
          type: string
          format: password
          example: super-secret-1
        client_type:
          type: string
          enum: [confidential, public]
          default: confidential
        token_endpoint_auth_method:
          type: string
          enum: [client_secret_basic, client_secret_post, none]
        require_pkce:
          type: boolean
          default: true
        require_consent:
          type: boolean
          default: true
        access_token_ttl_seconds:
          type: integer
          format: int32
          default: 3600
        refresh_token_ttl_seconds:
          type: integer
          format: int32
          default: 2592000
        id_token_ttl_seconds:
          type: integer
          format: int32
          default: 3600
        grant_types:
          type: array
          items:
            type: string
            enum: [authorization_code, refresh_token, client_credentials]
        scopes:
          type: array
          items:
            type: string
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
        post_logout_redirect_uris:
          type: array
          items:
            type: string
            format: uri
        status:
          type: string
          example: active
    CreateClientResponse:
      type: object
      required:
        - client_id
        - client_name
        - client_type
        - token_endpoint_auth_method
        - require_pkce
        - require_consent
        - access_token_ttl_seconds
        - refresh_token_ttl_seconds
        - id_token_ttl_seconds
        - grant_types
        - auth_methods
        - scopes
        - redirect_uris
        - status
      properties:
        client_id:
          type: string
        client_name:
          type: string
        client_type:
          type: string
        token_endpoint_auth_method:
          type: string
        require_pkce:
          type: boolean
        require_consent:
          type: boolean
        access_token_ttl_seconds:
          type: integer
          format: int32
        refresh_token_ttl_seconds:
          type: integer
          format: int32
        id_token_ttl_seconds:
          type: integer
          format: int32
        grant_types:
          type: array
          items:
            type: string
        auth_methods:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
        post_logout_redirect_uris:
          type: array
          items:
            type: string
            format: uri
        status:
          type: string
...